* Fixed serialization code to remove the "duplicate" symbols and removed function specialization, rely on standard function overloads instead, and define the functions as free functions with the object definitions instead * Cleaned up client application code.
