+++ blackjack/apps/server.cpp
@@ -4,6 +4,6 @@
 int main(int argc, char* argv[]) {
   Application app;
   app.setup(argc, argv);
-  app.run();
-  return 0;
+  int r = app.run();
+  return r;
 }
