+++ blackjack/include/player.h
@@ -12,6 +12,8 @@ struct player_t
   std::vector<int > hands; //< identifiers of the hands run by this player?
 };
 
+char* serialize(player_t& p, char* data);
+char* deserialize(char* buffer, player_t* p);
 
 
 #endif // BLACKJACK_PLAYER_H__
