+++ blackjack/include/serialize.h
@@ -20,32 +20,6 @@ char* serialize(std::vector<T>& list, char* data) {
   return data;
 }
 
-template <>
-char* serialize(hand_t& h, char* data) {
-  data = serialize(h.identifier, data);
-  data = serialize(h.state, data);
-  data = serialize(h.action, data);
-  data = serialize(h.cards, data);
-  return data;
-}
-
-template <>
-char* serialize(player_t& p, char* data) {
-  data = serialize(p.identifier, data);
-  data = serialize(p.hands, data);
-  return data;
-}
-
-template <>
-char* serialize(table_t& t, char* data) {
-  data = serialize(t.identifier, data);
-  data = serialize(t.state, data);
-  //data = serialize(t.deck, data); //< we dont actually want anyone to know whats contained here..?
-  data = serialize(t.players, data);
-  data = serialize(t.dealer, data);
-  return data;
-}
-
 
 
 template <class T>
@@ -63,34 +37,6 @@ char* deserialize(char* buffer, std::vector<T>* obj) {
     buffer = deserialize(buffer, &(*obj)[i]);
   }
 
-  return buffer;
-}
-
-template <>
-char* deserialize(char* buffer, hand_t* h) {
-  buffer = deserialize(buffer, &(h->identifier));
-  buffer = deserialize(buffer, &(h->state));
-  buffer = deserialize(buffer, &(h->action));
-  buffer = deserialize(buffer, &(h->cards));
-  return buffer;
-}
-
-template <>
-char* deserialize(char* buffer, player_t* p) {
-  buffer = deserialize(buffer, &(p->identifier));
-  buffer = deserialize(buffer, &(p->hands));
-  return buffer;
-}
-
-template <>
-char* deserialize(char* buffer, table_t* t) {
-  buffer = deserialize(buffer, &(t->identifier));
-  int state;
-  buffer = deserialize(buffer, &state);
-  (t->state) = (TableState)state;
-  //data = deserialize(buffer, &(t->deck)); //< we dont actually want anyone to know whats contained here..?
-  buffer = deserialize(buffer, &(t->players));
-  buffer = deserialize(buffer, &(t->dealer));
   return buffer;
 }
